requirements.txt
torch>=1.13.0
torchvision>=0.14.0
segmentation-models-pytorch>=0.2.1
albumentations>=1.3.0
timm>=0.6.12
numpy>=1.21.6
Pillow>=9.0.0
tqdm>=4.64.0
